refactor: rename composable from useGetGithubRepositories to useGithubRepositories
- Rename composable file from useGetGithubRepositories.ts to useGithubRepositories.ts
- Update import path in component to match new composable name
- Update test mock to use correct composable name and variable reference
- Standardize naming convention for consistency across codebase
